+++ xxcmd/main.py
@@ -4,7 +4,7 @@ import os
 from .cmdmanager import CmdManager
 
 
-__version__ = "0.7.0"
+__version__ = "0.8.0"
 
 
 def main():