+++ xxcmd/config.py
@@ -22,7 +22,10 @@ class Config():
             'whole-line-selection': True,
             'search-labels-only': False,
             'search-labels-first': True,
-            'shell': 'DEFAULT'
+            'shell': 'DEFAULT',
+            'sort-by-label': True,
+            'sort-by-command': False,
+            'sort-case-sensitive': True
         }
 
         # If there is a config file merge that in too