+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Added feature to edit existing commands with F2 or CTRL+I.
 - Added support for arrow keys, home, end, CTRL+arrow in line editor.
 
 ## [0.7.0] - 2021-02-09