+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added feature to edit existing commands with F2 or CTRL+I.
 - Added support for arrow keys, home, end, CTRL+arrow in line editor.
 - Added config file option to change shell used to execute commands.
+- Added sort options for interactive view. Defaults to sort by label.
 
 ## [0.7.0] - 2021-02-09
 - Changed default search to search only labels, if no matches try searching commands.